Renames replication related services.
refs #2059
