+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude`
-# on 2022-11-07 16:47:23 UTC using RuboCop version 1.38.0.
+# on 2022-11-23 15:19:45 UTC using RuboCop version 1.39.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -30,7 +30,7 @@ Lint/MissingSuper:
     - 'app/services/reporters/logger_reporter.rb'
     - 'config/initializers/okcomputer.rb'
 
-# Offense count: 22
+# Offense count: 21
 # Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, CountRepeatedAttributes, Max.
 Metrics/AbcSize:
   Exclude:
@@ -39,15 +39,15 @@ Metrics/AbcSize:
     - 'app/jobs/validate_moab_job.rb'
     - 'app/models/replication/druid_version_zip.rb'
     - 'app/models/preserved_object.rb'
-    - 'app/services/audit/catalog_to_archive.rb'
+    - 'app/services/audit/replication_to_endpoint_base.rb'
     - 'app/services/audit/catalog_to_moab.rb'
-    - 'app/services/audit/checksum_validator_utils.rb'
-    - 'app/services/catalog_utils.rb'
     - 'app/services/audit/checksum_validator.rb'
+    - 'app/services/audit/checksum_validator_utils.rb'
+    - 'app/services/audit/replication_support.rb'
+    - 'app/services/catalog_utils.rb'
     - 'app/services/complete_moab_service/check_existence.rb'
     - 'app/services/complete_moab_service/update_version.rb'
     - 'app/services/moab_storage_root_report_service.rb'
-    - 'app/services/s3/s3_audit.rb'
     - 'config/initializers/okcomputer.rb'
 
 # Offense count: 2
@@ -57,7 +57,7 @@ Metrics/ClassLength:
     - 'app/controllers/objects_controller.rb'
     - 'app/models/replication/druid_version_zip.rb'
 
-# Offense count: 37
+# Offense count: 36
 # Configuration parameters: CountComments, CountAsOne, ExcludedMethods, AllowedMethods, AllowedPatterns, IgnoredMethods.
 Metrics/MethodLength:
   Max: 33
@@ -95,10 +95,10 @@ RSpec/StubbedMock:
     - 'spec/jobs/part_replication_audit_job_spec.rb'
     - 'spec/services/audit/catalog_to_moab_spec.rb'
     - 'spec/services/audit/checksum_validator_spec.rb'
+    - 'spec/services/audit/shared_examples_replication_to_endpoint.rb'
     - 'spec/services/complete_moab_service/check_existence_spec.rb'
     - 'spec/services/complete_moab_service/create_after_validation_spec.rb'
     - 'spec/services/complete_moab_service/update_version_after_validation_spec.rb'
-    - 'spec/services/s3/shared_examples_s3_audit.rb'
 
 # Offense count: 4
 RSpec/SubjectDeclaration:
@@ -108,7 +108,7 @@ RSpec/SubjectDeclaration:
     - 'spec/services/reporters/honeybadger_reporter_spec.rb'
     - 'spec/services/reporters/logger_reporter_spec.rb'
 
-# Offense count: 18
+# Offense count: 19
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: slashes, arguments
 Rails/FilePath:
@@ -156,14 +156,13 @@ Rails/TimeZone:
 # Offense count: 6
 Style/CombinableLoops:
   Exclude:
-    - 'spec/services/moab_on_storage/storage_directory_spec.rb'
     - 'spec/services/audit/catalog_to_moab_spec.rb'
+    - 'spec/services/moab_on_storage/storage_directory_spec.rb'
 
-# Offense count: 5
+# Offense count: 4
 # Configuration parameters: AllowedConstants.
 Style/Documentation:
   Exclude:
-    - 'app/controllers/application_controller.rb'
     - 'app/helpers/application_helper.rb'
     - 'app/models/application_record.rb'
     - 'config/application.rb'
@@ -173,7 +172,7 @@ Style/Documentation:
 # AllowedMethods: respond_to_missing?
 Style/OptionalBooleanParameter:
   Exclude:
-    - 'app/services/s3/s3_audit.rb'
+    - 'app/services/audit/replication_to_endpoint_base.rb'
 
 # Offense count: 4
 # This cop supports unsafe autocorrection (--autocorrect-all).
