+++ app/services/audit/replication_support.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Audit
-  # Catalog to cloud archive provider (what ZipEndpoint, ZippedMoabVersion etc represent) audit code
-  class CatalogToArchive
+  # Methods to support auditing replication.
+  class ReplicationSupport
     def self.logger
       @logger ||= Logger.new(Rails.root.join('log', 'c2a.log'))
     end
