+++ app/services/replication/aws_provider.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-module S3
-  # The Application's configured interface to IBM's S3 compatible service.
-  class IbmProvider
+module Replication
+  # The Application's configured interface to AWS S3.
+  class AwsProvider
     delegate :client, to: :resource
     attr_reader :aws_client_args
 
@@ -10,8 +10,7 @@ module S3
       @aws_client_args = {
         region: region,
         access_key_id: access_key_id,
-        secret_access_key: secret_access_key,
-        endpoint: Settings.zip_endpoints.ibm_us_south.endpoint_node
+        secret_access_key: secret_access_key
       }
     end
 
@@ -22,12 +21,19 @@ module S3
 
     # @return [String]
     def bucket_name
-      Settings.zip_endpoints.ibm_us_south.storage_location
+      Settings.zip_endpoints[region_config_section].storage_location
     end
 
     # @return [::Aws::S3::Resource]
     def resource
       ::Aws::S3::Resource.new(aws_client_args)
+    end
+
+    # Allow this class to dynamically pull back config for different AWS
+    # regions. Given a region of "us-west-2", returns "aws_s3_west_2"
+    def region_config_section
+      region_parts = aws_client_args[:region].split('-')
+      "aws_s3_#{region_parts[1]}_#{region_parts[2]}"
     end
   end
 end
