+++ config/settings/test.yml
@@ -20,7 +20,7 @@ zip_endpoints:
     endpoint_node: 'us-west-2'
     storage_location: 'sul-sdr-aws-us-west-2-test'
     delivery_class: 'S3WestDeliveryJob'
-    audit_class: 'S3::Aws::Audit'
+    audit_class: 'Audit::ReplicationToAws'
     access_key_id: 'overridden-by-env-var-in-ci'
     secret_access_key: 'overridden-by-env-var-in-ci'
   ibm_us_south:
@@ -28,7 +28,7 @@ zip_endpoints:
     endpoint_node: 'https://s3.us-south.cloud-object-storage.appdomain.cloud'
     storage_location: 'sul-sdr-ibm-us-south-1-test'
     delivery_class: 'IbmSouthDeliveryJob'
-    audit_class: 'S3::Ibm::Audit'
+    audit_class: 'Audit::ReplicationToIbm'
     access_key_id: 'overridden-by-env-var-in-ci'
     secret_access_key: 'overridden-by-env-var-in-ci'
