+++ spec/services/replication/aws_provider_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-require 'services/s3/shared_examples_provider'
+require 'services/replication/shared_examples_provider'
 
-describe S3::AwsProvider do
+describe Replication::AwsProvider do
   it_behaves_like 'provider', described_class,
                   Settings.zip_endpoints.aws_s3_west_2.storage_location,
                   Settings.zip_endpoints.aws_s3_west_2.region,
